+++ R/rand_test.R
@@ -2,23 +2,25 @@
 #'
 #' The `rand_test` function computes a randomization test for single or multiple
 #' baseline single-case data.  The function is based on an algorithm from the
-#' `SCRT` package (Bulte & Onghena, 2009, 2012), but rewritten and extended
-#' for the use in AB designs.
+#' `SCRT` package (Bulte & Onghena, 2009, 2012), but rewritten and extended for
+#' the use in AB designs.
 #'
 #'
 #' @inheritParams .inheritParams
 #' @param statistic Defines the statistic on which the comparison of phases A
 #'   and B is based on. Default setting is `statistic = "Mean B-A"`. The
-#' following comparisons are possible: \itemize{ \item\code{"Mean A-B"}: Uses
+#' following comparisons are possible: \itemize{ \item`Mean A-B`: Uses
 #' the difference between the mean of phase A and the mean of phase B. This is
 #' appropriate if a decrease of scores was expected for phase B.
-#' \item\code{"Mean B-A"}: Uses the difference between the mean of phase B and
+#' \item`Mean B-A`: Uses the difference between the mean of phase B and
 #' the mean of phase A. This is appropriate if an increase of scores was
-#' expected for phase B.  \item\code{"Mean |A-B|"}: Uses the absolute value of
-#' the difference between the means of phases A and B.  \item\code{"Median
-#' A-B"}: The same as \code{"Mean A-B"}, but based on the median.
-#' \item\code{"Median B-A"}: The same as \code{"Mean B-A"}, but based on the
-#' median.  }
+#' expected for phase B.  \item`Mean |A-B|`: Uses the absolute value of
+#' the difference between the means of phases A and B.  \item`Median
+#' A-B`: The same as `Mean A-B`, but based on the median.
+#' \item`Median B-A`: The same as `Mean B-A`, but based on the
+#' median.  \item`SMD hedges / SMD glass`: Standardizes mean difference of B-A 
+#' as Hedges's g or Glass' delta. \item`NAP`: Non-overlap of all pairs.
+#' \item`W-test`: Wilcoxon-test statistic W.\item`T-test`: T-test statistic t.}
 #' @param number Sample size of the randomization distribution. The exactness of
 #'   the p-value can not exceed \eqn{1/number} (i.e., `number = 100` results in
 #'   p-values with an exactness of one percent). Default is `number = 500`. For
@@ -53,7 +55,7 @@
 #'   points in phase B.} \item{limit}{Numeric from function call (see
 #'   \code{Arguments} above).}
 #' \item{startpoints}{A vector defining the start points passed from the
-#' function call (see \code{Arguments} above).} \item{p.value}{P-value of the
+#' function call (see `Arguments` above).} \item{p.value}{P-value of the
 #' randomization test for the given data.} \item{number}{Sample size of
 #' randomization distribution from function call (see \code{Arguments} above).}
 #' \item{complete}{Logical argument from function call (see \code{Arguments}
@@ -62,10 +64,11 @@
 #' \item{Z}{Z-value of observed test statistic.} \item{p.z.single}{Probability
 #' of z-value.} \item{distribution}{Test statistic distribution from randomized
 #' data sets.} \item{possible.combinations}{Number of possible combinations
-#' under the given restrictions.} \item{auto.corrected.number}{\code{TRUE}
+#' under the given restrictions.} \item{auto.corrected.number}{`TRUE`
 #' indicates that a corrected number of combinations was used. This happens, if
 #' the number of possible combinations (under the given restrictions) undercuts
-#' the requested \code{number} of combinations.}
+#' the requested `number` of combinations.} \item{ecxlude.equal}{see argument
+#' above}
 #' @author Juergen Wilbert
 #' @references Bulte, I., & Onghena, P. (2009). Randomization tests for
 #'   multiple-baseline designs: An extension of the SCRT-R package.
@@ -84,7 +87,10 @@
 #' @export
 rand_test <- function (data, dvar, pvar, 
                        statistic = c("Mean B-A", "Mean A-B", "Median B-A", 
-                                     "Median A-B", "Mean |A-B|", "Median |A-B|"), 
+                                     "Median A-B", "Mean |A-B|", "Median |A-B|",
+                                     "SMD hedges", "SMD glass", 
+                                     "W-test", "T-test", 
+                                     "NAP", "NAP decreasing"), 
                        number = 500, 
                        complete = FALSE, 
                        limit = 5, 
@@ -119,6 +125,8 @@ rand_test <- function (data, dvar, pvar,
   obs <- lapply(data, function(x) x[, dvar])
   MT  <- lapply(data, nrow)
   N   <- length(data)
+  
+  testdirection <- "greater"
   
   if (identical(exclude.equal, "auto")) exclude.equal <- N == 1
   
@@ -155,123 +163,129 @@ rand_test <- function (data, dvar, pvar,
   
 # Sample Random A and B phases ---------------------------------------------
   
-  rnd.a <- list()
+  rnd_a <- list()
   for (i in 1:number) {
     ascores <- list()
     for (case in 1:N)
       ascores[[case]] <- data[[case]][1:(startpts[i, case] - 1), dvar]
-    rnd.a[[i]] <- ascores
+    rnd_a[[i]] <- ascores
   }
   
-  rnd.b <- list()
+  rnd_b <- list()
   for (i in 1:number) {
     ascores <- list()
     for (case in 1:N)
       ascores[[case]] <- data[[case]][startpts[i, case]:MT[[case]], dvar]
-    rnd.b[[i]] <- ascores
+    rnd_b[[i]] <- ascores
   }
   
 # Functions for phase differences -----------------------------------------
 
-  if (statistic == "B-A" || statistic == "Mean B-A") {
-    means.b <- unlist(lapply(rnd.b, function(x) lapply(x, mean,na.rm = TRUE)))
-    means.a <- unlist(lapply(rnd.a, function(x) lapply(x, mean,na.rm = TRUE)))
-    ma <- matrix(means.b - means.a, ncol = N, nrow = number, byrow = TRUE)
-    
-    dist <- apply(ma, 1, mean, na.rm = TRUE)
-    means.b <- unlist(lapply(b, mean,na.rm = TRUE))
-    means.a <- unlist(lapply(a, mean,na.rm = TRUE))
-    ma <- matrix(means.b - means.a, ncol = N, nrow = 1, byrow = TRUE)
-    obs.stat <- mean(ma,na.rm = TRUE)
-  }
+  if (statistic == "SMD hedges") {
+    res <- rand_test_smd(rnd_a, rnd_b, a, b, method = "hedges")
+  }  
   
-  if (statistic == "A-B" || statistic == "Mean A-B") {
-    means.b <- unlist(lapply(rnd.b, function(x) lapply(x, mean,na.rm = TRUE)))
-    means.a <- unlist(lapply(rnd.a, function(x) lapply(x, mean,na.rm = TRUE)))
-    ma <- matrix(means.a - means.b, ncol = N, nrow = number, byrow = TRUE)
-    
-    dist <- apply(ma, 1, mean,na.rm = TRUE)
-    means.b <- unlist(lapply(b, mean, na.rm = TRUE))
-    means.a <- unlist(lapply(a, mean, na.rm = TRUE))
-    ma <- matrix(means.a - means.b, ncol = N, nrow = 1, byrow = TRUE)
-    obs.stat <- mean(ma,na.rm = TRUE)
-  }
+  if (statistic == "SMD glass") {
+    res <- rand_test_smd(rnd_a, rnd_b, a, b, method = "glass")
+  }  
   
-  if (statistic == "Median B-A") {
-    medians.b <- unlist(lapply(rnd.b, function(x) lapply(x,median,na.rm = TRUE)))
-    medians.a <- unlist(lapply(rnd.a, function(x) lapply(x,median,na.rm = TRUE)))
-    ma <- matrix(medians.b-medians.a, ncol = N, nrow = number, byrow = TRUE)
-    
-    dist <- apply(ma,1,median,na.rm = TRUE)
-    medians.b <- unlist(lapply(b, median,na.rm = TRUE))
-    medians.a <- unlist(lapply(a, median,na.rm = TRUE))
-    ma <- matrix(medians.b-medians.a, ncol = N, nrow = 1, byrow = TRUE)
-    obs.stat <- median(ma,na.rm = TRUE)
-  }
+  if (statistic == "W-test") {
+    res <- rand_test_u_test(rnd_a, rnd_b, a, b)
+    testdirection <- "less"
+  }   
+
+  if (statistic == "T-test") {
+    res <- rand_test_t_test(rnd_a, rnd_b, a, b)
+  } 
   
-  if (statistic == "Median A-B") {
-    medians.b <- unlist(lapply(rnd.b, function(x) lapply(x, median, na.rm = TRUE)))
-    medians.a <- unlist(lapply(rnd.a, function(x) lapply(x, median, na.rm = TRUE)))
-    ma <- matrix(medians.a - medians.b, ncol = N, nrow = number, byrow = TRUE)
-    
-    dist <- apply(ma, 1, median,na.rm = TRUE)
-    medians.b <- unlist(lapply(b, median, na.rm = TRUE))
-    medians.a <- unlist(lapply(a, median, na.rm = TRUE))
-    ma <- matrix(medians.a - medians.b, ncol = N, nrow = 1, byrow = TRUE)
-    obs.stat <- median(ma, na.rm = TRUE)
-  }	
+  if (statistic == "NAP") {
+    res <- rand_test_nap(rnd_a, rnd_b, a, b)
+  } 
   
+  if (statistic == "NAP decreasing") {
+    res <- rand_test_nap(rnd_a, rnd_b, a, b, decreasing = TRUE)
+  } 
+  
+  if (statistic == "Mean B-A") {
+    res <- rand_test_avg(
+      rnd_a, rnd_b, a, b, 
+      fn = function(x) mean(x, na.rm = TRUE),
+      method = "B-A"
+    )
+   
+  } 
+  
+  if (statistic == "Mean A-B") {
+    res <- rand_test_avg(
+      rnd_a, rnd_b, a, b, 
+      fn = function(x) mean(x, na.rm = TRUE),
+      method = "A-B"
+    )
+  } 
   
   if (statistic == "Mean |A-B|") {
-    means.b <- unlist(lapply(rnd.b, function(x) lapply(x, mean, na.rm = TRUE)))
-    means.a <- unlist(lapply(rnd.a, function(x) lapply(x, mean, na.rm = TRUE)))
-    ma <- matrix(abs(means.a - means.b), ncol = N, nrow = number, byrow = TRUE)
-    
-    dist <- apply(ma, 1, mean,na.rm = TRUE)
-    means.b <- unlist(lapply(b, mean, na.rm = TRUE))
-    means.a <- unlist(lapply(a, mean, na.rm = TRUE))
-    ma <- matrix(abs(means.a - means.b), ncol = N, nrow = 1, byrow = TRUE)
-    obs.stat <- mean(ma, na.rm = TRUE)
-  }
+    res <- rand_test_avg(
+      rnd_a, rnd_b, a, b, 
+      fn = function(x) mean(x, na.rm = TRUE),
+      method = "abs"
+    )
+  } 
 
+  if (statistic == "Median B-A") {
+    res <- rand_test_avg(
+      rnd_a, rnd_b, a, b, 
+      fn = function(x) median(x, na.rm = TRUE),
+      method = "B-A"
+    )
+  } 
+  
+  if (statistic == "Median A-B") {
+    res <- rand_test_avg(
+      rnd_a, rnd_b, a, b, 
+      fn = function(x) median(x, na.rm = TRUE),
+      method = "A-B"
+    )
+  } 
+  
   if (statistic == "Median |A-B|") {
-    medians.b <- unlist(lapply(rnd.b, function(x) lapply(x, median, na.rm = TRUE)))
-    medians.a <- unlist(lapply(rnd.a, function(x) lapply(x, median, na.rm = TRUE)))
-    ma <- matrix(abs(medians.a - medians.b), ncol = N, nrow = number, byrow = TRUE)
-    dist <- apply(ma, 1, median,na.rm = TRUE)
-    
-    medians.b <- unlist(lapply(b, median, na.rm = TRUE))
-    medians.a <- unlist(lapply(a, median, na.rm = TRUE))
-    ma <- matrix(abs(medians.a - medians.b), ncol = N, nrow = 1, byrow = TRUE)
-    obs.stat <- median(ma, na.rm = TRUE)
-  }
-
+    res <- rand_test_avg(
+      rnd_a, rnd_b, a, b, 
+      fn = function(x) median(x, na.rm = TRUE),
+      method = "abs"
+    )
+  }   
+  
 # p value -----------------------------------------------------------------
 
-  if (!exclude.equal) test <- dist >= obs.stat else test <- dist > obs.stat
+  if (testdirection == "greater") {
+    test <- if (!exclude.equal) res$dist >= res$obs_stat else res$dist > res$obs_stat
+  } else {
+    test <- if (!exclude.equal) res$dist <= res$obs_stat else res$dist < res$obs_stat
+  }
   
   p.value <- sum(test) / number
   
 # return ------------------------------------------------------------------
 
   if (graph){
-    h <- hist(dist, plot = FALSE)
-    lab <- paste0(round(h$counts / length(dist) * 100, 0), "%")
+    h <- hist(res$dist, plot = FALSE)
+    lab <- paste0(round(h$counts / length(res$dist) * 100, 0), "%")
     xlim <- c(min(h$breaks,na.rm = TRUE), max(h$breaks, na.rm = TRUE))
     ylim <- round(max(h$counts * 1.2))
-    if (obs.stat < xlim[1]) xlim[1] <- obs.stat
-    if (obs.stat > xlim[2]) xlim[2] <- obs.stat
+    if (res$obs_stat < xlim[1]) xlim[1] <- res$obs_stat
+    if (res$obs_stat > xlim[2]) xlim[2] <- res$obs_stat
     hist(
-      dist, xlab = statistic, labels = lab, xlim = xlim, ylim = c(0, ylim), 
+      res$dist, xlab = statistic, labels = lab, xlim = xlim, ylim = c(0, ylim), 
       ylab = "Frequency", main = "Random distribution", col = "lightgrey"
     )
-    abline(v = obs.stat, lty = 2, lwd = 2, col = "grey") 
+    abline(v = res$obs_stat, lty = 2, lwd = 2, col = "grey") 
     if (p.value < 0.5) pos <- 2 else pos <- 4
-    text(obs.stat, ylim, "observed", pos = pos)
+    text(res$obs_stat, ylim, "observed", pos = pos)
   }
   
-  Z <- (obs.stat - mean(dist, na.rm = TRUE)) / sd(dist, na.rm = TRUE)
-  p.Z.single <- 1 - pnorm(Z)
+  Z <- (res$obs_stat - mean(res$dist, na.rm = TRUE)) / sd(res$dist, na.rm = TRUE)
+  p.Z.single <- if (testdirection == "greater") 1 - pnorm(Z) else pnorm(Z)
+    
   possible.combinations <- cumprod(unlist(lapply(pos.startpts, length)))[N]
 
   out <- list(
@@ -286,12 +300,14 @@ rand_test <- function (data, dvar, pvar,
     p.value = p.value, 
     number = number, 
     complete = complete, 
-    observed.statistic = obs.stat, 
+    observed.statistic = res$obs_stat, 
     Z = Z, 
     p.Z.single = p.Z.single, 
-    distribution = dist, 
+    distribution = res$dist, 
     possible.combinations = possible.combinations, 
-    auto.corrected.number = auto.corrected.number
+    auto.corrected.number = auto.corrected.number,
+    exclude.equal = exclude.equal,
+    testdirection = testdirection
   )
   
   class(out) <- c("sc_rand")
@@ -302,18 +318,225 @@ rand_test <- function (data, dvar, pvar,
 
 }
 
-#' @rdname deprecated-functions
-#' @export
-randSC <- function(...) {
-  .deprecated_warning("rand_test", "randSC")
-  rand_test(...)
+rand_test_avg <- function(rnd_a, rnd_b, a, b, fn, method) {
+  
+  N <- length(rnd_a[[1]])
+  number <- length(rnd_a)
+  avg_b <- unlist(lapply(rnd_b, function(x) lapply(x, fn)))
+  avg_a <- unlist(lapply(rnd_a, function(x) lapply(x, fn)))
+  
+  ma <- if (method == "B-A") {
+    matrix(avg_b - avg_a, ncol = N, nrow = number, byrow = TRUE)
+  } else if (method == "A-B") {
+    matrix(avg_a - avg_b, ncol = N, nrow = number, byrow = TRUE)
+  } else if (method == "abs") {
+    matrix(abs(avg_a - avg_b), ncol = N, nrow = number, byrow = TRUE)
+  }
+  
+  dist <- apply(ma, 1, fn)
+  
+  avg_b <- unlist(lapply(b, fn))
+  avg_a <- unlist(lapply(a, fn))
+  
+  if (method == "B-A") {
+    obs_stat <- avg_b - avg_a
+  } else if (method == "A-B") {
+    obs_stat <- avg_a - avg_b
+  } else if (method == "abs") {
+    obs_stat <- abs(avg_a - avg_b)
+  }
+  
+  list(
+    obs_stat = fn(obs_stat),
+    dist = dist
+  )
+  
 }
 
-#' @rdname deprecated-functions
-#' @export
-rand.test <- function(...) {
-  .deprecated_warning("rand_test", "rand.test")
-  rand_test(...)
+
+rand_test_smd <- function(rnd_a, rnd_b, a, b, method) {
+  
+  
+  N <- length(rnd_a[[1]])
+  number <- length(rnd_a)
+  
+  stats_a <- unlist(lapply(rnd_a, function(x) lapply(x, mean, na.rm = TRUE)))
+  stats_b <- unlist(lapply(rnd_b, function(x) lapply(x, mean, na.rm = TRUE)))
+  sds_a <- unlist(lapply(rnd_a, function(x) lapply(x, sd, na.rm = TRUE)))
+  sds_b <- unlist(lapply(rnd_b, function(x) lapply(x, sd, na.rm = TRUE)))
+  ns_a <- unlist(lapply(rnd_a, function(x) lapply(x, length)))
+  ns_b <- unlist(lapply(rnd_b, function(x) lapply(x, length)))
+  
+  # Hedges'g + Durlak correction or Glass' Delta
+  
+  if (method == "hedges") {
+    dat <- (stats_b - stats_a) / 
+           sqrt(((ns_a - 1) * sds_a^2 + (ns_b - 1) * sds_b^2) / (ns_a + ns_b - 2)) *
+           (((ns_a+ns_b) - 3) / ((ns_a+ns_b) - 2.25) * sqrt(((ns_a+ns_b) - 2) / (ns_a+ns_b)))
+  } else if (method == "glass") {
+    dat <- (stats_b - stats_a) / sds_a
+  }
+  
+  ma <- matrix(
+    dat, 
+    ncol = N, 
+    nrow = number, 
+    byrow = TRUE
+  )   
+ 
+  dist <- apply(ma, 1, function(x) sqrt(sum(x^2) / length(x)))
+  
+  stats_a <- unlist(lapply(a, mean, na.rm = TRUE))
+  stats_b <- unlist(lapply(b, mean, na.rm = TRUE))
+  sds_a <- unlist(lapply(a, sd, na.rm = TRUE))
+  sds_b <- unlist(lapply(b, sd, na.rm = TRUE)) 
+  ns_a <- unlist(lapply(a, length))
+  ns_b <- unlist(lapply(b, length))
+  
+  if (method == "hedges") {
+    dat <- (stats_b - stats_a) / 
+      sqrt(((ns_a - 1) * sds_a^2 + (ns_b - 1) * sds_b^2) / (ns_a + ns_b - 2)) *
+      (((ns_a+ns_b) - 3) / ((ns_a+ns_b) - 2.25) * sqrt(((ns_a+ns_b) - 2) / (ns_a+ns_b)))
+  } else if (method == "glass") {
+    dat <- (stats_b - stats_a) / sds_a
+  }
+
+  list(
+    obs_stat = sqrt(sum(dat^2) / length(dat)),
+    dist = dist
+  )
 }
 
+rand_test_t_test <- function(rnd_a, rnd_b, a, b) {
+  
+  
+  N <- length(rnd_a[[1]])
+  number <- length(rnd_a)
+ 
+  rnd_a <- unlist(rnd_a,recursive = FALSE)
+  rnd_b <- unlist(rnd_b,recursive = FALSE)
+  
+  dat <- mapply(
+    function(a, b) {
+      suppressWarnings(t.test(b,a)$statistic)
+    }, 
+    a = rnd_a, b = rnd_b
+  )
+  
+  ma <- matrix(
+    dat, 
+    ncol = N, 
+    nrow = number, 
+    byrow = TRUE
+  )   
+  
+  dist <- apply(ma, 1, mean, na.rm = TRUE)
+  
+  dat <- mapply(
+    function(a, b) {
+      suppressWarnings(t.test(b,a)$statistic)
+    }, 
+    a = a, b = b
+  )
+  
+  list(
+    obs_stat = mean(dat, na.rm = TRUE),
+    dist = dist
+  )
+}
 
+rand_test_u_test <- function(rnd_a, rnd_b, a, b) {
+  
+  
+  N <- length(rnd_a[[1]])
+  number <- length(rnd_a)
+  
+  rnd_a <- unlist(rnd_a,recursive = FALSE)
+  rnd_b <- unlist(rnd_b,recursive = FALSE)
+  
+  dat <- mapply(
+    function(a, b) {
+      suppressWarnings(wilcox.test(a,b)$statistic)
+    }, 
+    a = rnd_a, b = rnd_b
+  )
+  
+  ma <- matrix(
+    dat, 
+    ncol = N, 
+    nrow = number, 
+    byrow = TRUE
+  )   
+  
+  dist <- apply(ma, 1, mean, na.rm = TRUE)
+  
+  dat <- mapply(
+    function(a, b) {
+      suppressWarnings(wilcox.test(a,b)$statistic)
+    }, 
+    a = a, b = b
+  )
+  
+  list(
+    obs_stat = mean(dat, na.rm = TRUE),
+    dist = dist
+  )
+}
+
+rand_test_nap <- function(rnd_a, rnd_b, a, b, decreasing = FALSE) {
+ 
+  N <- length(rnd_a[[1]])
+  number <- length(rnd_a)
+  rnd_a <- unlist(rnd_a,recursive = FALSE)
+  rnd_b <- unlist(rnd_b,recursive = FALSE)
+
+  dat <- mapply(
+    function(a, b) {
+     
+      if (!decreasing) {
+        pos <- sum(unlist(lapply(a, function(x) b > x)))
+      }
+      if (decreasing) {
+        pos <- sum(unlist(lapply(a, function(x) b < x)))
+      }
+      
+      ties <- sum(unlist(lapply(a, function(x) x == b)))
+      non_overlaps <- pos + (0.5 * ties)
+      pairs <- length(a) * length(b)
+      nap <- non_overlaps / pairs * 100
+    }, 
+    a = rnd_a, b = rnd_b
+  )
+  
+  ma <- matrix(
+    dat, 
+    ncol = N, 
+    nrow = number, 
+    byrow = TRUE
+  )   
+  
+  dist <- apply(ma, 1, mean, na.rm = TRUE)
+  
+  dat <- mapply(
+    function(a, b) {
+      if (!decreasing) {
+        pos <- sum(unlist(lapply(a, function(x) b > x)))
+      }
+      if (decreasing) {
+        pos <- sum(unlist(lapply(a, function(x) b < x)))
+      }
+    
+      ties <- sum(unlist(lapply(a, function(x) x == b)))
+      non_overlaps <- pos + (0.5 * ties)
+      pairs <- length(a) * length(b)
+      nap <- non_overlaps / pairs * 100
+    }, 
+    a = a, b = b
+  )
+  
+  list(
+    obs_stat = mean(dat, na.rm = TRUE),
+    dist = dist
+  )
+
+}