+++ R/scan.R
@@ -1,5 +1,11 @@
 .onAttach <- function(lib, pkg, ...) {
-  packageStartupMessage(opt("startup_message"))
+  packageStartupMessage(
+    "\033[34m", 
+    opt("startup_message"),
+    "\033[31m",
+    opt("tip")[[sample(1:length(opt("tip")), 1)]]
+  )
+  
 }	
 
 .onLoad <- function(lib, pkg, ...) {